refactor: address code review feedback (6 items)
1. Extract LocaleDetection concern to eliminate parse_accept_language
   duplication across ApplicationController, Madmin::ApplicationController,
   and Admins::SessionsController
2. Unify detect_browser_locale into the shared concern, remove
   duplicate implementations
3. Use includes(:user) instead of find_each in
   broadcast_to_other_participants to avoid N+1 user loads
4. Simplify conversation message partial — current_user is always
   available now (from controller or broadcast locals)
5. Inline translation_locales query in Article#embedding_source_text,
   removing unnecessary private method
6. Add should_reembed? check in EmbedRecordJob#perform to skip
   duplicate embeds from concurrent TranslateContentJob workers
