+++ app/controllers/application_controller.rb
@@ -1,4 +1,6 @@
 class ApplicationController < ActionController::Base
+  include LocaleDetection
+
   rate_limit to: 100, within: 1.minute, name: "global"
 
   # Only allow modern browsers supporting webp images, web push, badges, import maps, CSS nesting, and CSS :has.
@@ -32,14 +34,8 @@ class ApplicationController < ActionController::Base
     end
 
     # 2. Accept-Language header
-    if request.headers["Accept-Language"]
-      accepted = parse_accept_language(request.headers["Accept-Language"])
-      enabled = Language.enabled_codes
-
-      accepted.each do |code|
-        return code.to_sym if enabled.include?(code)
-      end
-    end
+    detected = detect_browser_locale
+    return detected.to_sym if detected
 
     # 3. Platform default language
     Setting.default_language.to_sym
@@ -182,14 +178,5 @@ class ApplicationController < ActionController::Base
     return if Setting.ai_chats_enabled?
 
     redirect_to team_root_path(current_team), alert: t("controllers.application.chats_disabled")
-  end
-
-  def parse_accept_language(header)
-    header.to_s.split(",").filter_map { |entry|
-      lang, quality = entry.strip.split(";")
-      code = lang&.strip&.split("-")&.first&.downcase
-      q = quality ? quality.strip.delete_prefix("q=").to_f : 1.0
-      [ code, q ] if code.present?
-    }.sort_by { |_, q| -q }.map(&:first).uniq
   end
 end
